Updated productsManager so it forwards the products calls to fetchers for sk1 and sk2
