+++ Purchases/Purchasing/StoreKit2/ProductsFetcherSK2.swift
@@ -1,0 +1,41 @@
+//
+//  Copyright RevenueCat Inc. All Rights Reserved.
+//
+//  Licensed under the MIT License (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//      https://opensource.org/licenses/MIT
+//
+//  ProductsManagerSK2.swift
+//
+//  Created by Andrés Boedo on 7/23/21.
+
+import Foundation
+import StoreKit
+
+enum ProductsManagerSK2Error: Error {
+
+    case productsRequestError(innerError: Error)
+
+}
+
+struct ProductsFetcherSK2 {
+
+    func products(identifiers: Set<String>) async throws -> Set<ProductWrapper> {
+        if #available(iOS 15.0, tvOS 15.0, macOS 13.0, watchOS 7.0, *) {
+
+            do {
+                let storeKitProducts = try await StoreKit.Product.products(for: identifiers)
+                let sk2Wrappers = storeKitProducts.map { SK2ProductWrapper(sk2Product: $0) }
+                return Set(sk2Wrappers)
+            } catch let error {
+                throw ProductsManagerSK2Error.productsRequestError(innerError: error)
+            }
+
+        } else {
+            return Set()
+        }
+    }
+
+}
